feat: add support for `backgroundType` parameter
